Split execution of expressions into //compiler and //executor
