+++ compiler/instruction/expression/mod.rs
@@ -9,9 +9,30 @@ use std::{
     fmt::{Debug, Display, Formatter},
 };
 
+use encoding::value::value::DBValue;
+use encoding::value::value_type::ValueTypeCategory;
+
+use crate::{
+    expression::expression_compiler::ExpressionCompilationContext,
+    inference::ExpressionCompilationError,
+    instruction::expression::op_codes::ExpressionOpCode,
+};
+use crate::instruction::expression::builtins::binary::BinaryExpression;
+use crate::instruction::expression::builtins::load_cast::ImplicitCast;
+use crate::instruction::expression::builtins::unary::UnaryExpression;
+
 pub mod builtins;
-pub mod evaluator;
 pub mod op_codes;
+
+pub trait ExpressionInstruction: Sized {
+    const OP_CODE: ExpressionOpCode;
+}
+
+pub trait CompilableExpression: ExpressionInstruction {
+    fn return_value_category(&self) -> Option<ValueTypeCategory>;
+
+    fn validate_and_append(builder: &mut ExpressionCompilationContext<'_>) -> Result<(), ExpressionCompilationError>;
+}
 
 pub enum ExpressionEvaluationError {
     CheckedOperationFailed,
