+++ compiler/tests/BUILD
@@ -6,21 +6,6 @@ load("@vaticle_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
 load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_test")
 package(default_visibility = ["//visibility:public",])
 
-rust_test(
-    name = "expressions",
-    crate_root = "expressions.rs",
-    srcs =  ["expressions.rs"],
-    deps = [
-        "//answer",
-        "//common/bytes",
-        "//compiler",
-        "//encoding",
-        "//ir",
-        "@vaticle_typeql//rust:typeql",
-        "@crates//:itertools",
-    ],
-)
-
 checkstyle_test(
     name = "checkstyle",
     include = glob(["*"]),
