+++ executor/lib.rs
@@ -17,6 +17,7 @@ mod function_executor;
 pub(crate) mod instruction;
 pub mod pattern_executor;
 pub mod program_executor;
+pub mod expression_executor;
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash, Ord, PartialOrd)]
 pub struct VariablePosition {
